+++ tests/03_model_test.js
@@ -1,10 +1,13 @@
 let user_utils = require('./user_utils');
 let model_building_utils = require('./model_building_utils');
+let mouse_tracking = require('./mouse_tracking.js');
 
 module.exports = {
 
     beforeEach : function (client) {
         client.url('http://localhost:8124/atompm').pause(300);
+
+        mouse_tracking.track_mouse(client);
     },
 
     'Login' : function (client) {
