+++ tests/04_compile_test.js
@@ -1,5 +1,6 @@
 let user_utils = require('./user_utils');
 let model_building_utils = require('./model_building_utils');
+let mouse_tracking = require('./mouse_tracking.js');
 
 let folder_name = "Formalisms/__LanguageSyntax__/SimpleClassDiagram";
 
@@ -7,6 +8,8 @@ module.exports = {
 
     beforeEach : function (client) {
         client.url('http://localhost:8124/atompm').pause(300);
+
+        mouse_tracking.track_mouse(client);
     },
 
     'Login' : function (client) {
