+++ tests/07_transformation_test.js
@@ -2,6 +2,7 @@
 
 let user_utils = require('./user_utils');
 let model_building_utils = require('./model_building_utils');
+let mouse_tracking = require('./mouse_tracking.js');
 
 let rule_toolbars = [
     "/Formalisms/__Transformations__/TransformationRule/TransformationRule.defaultIcons.metamodel",
@@ -12,6 +13,8 @@ module.exports = {
 
     beforeEach: function (client, done) {
         client.url('http://localhost:8124/atompm').pause(300).maximizeWindow(done);
+
+        mouse_tracking.track_mouse(client);
     },
 
     'Login': function (client) {
