+++ tests/08_pacman_transformation_test.js
@@ -2,11 +2,14 @@
 
 let user_utils = require('./user_utils');
 let model_building_utils = require('./model_building_utils');
+let mouse_tracking = require('./mouse_tracking.js');
 
 module.exports = {
 
     beforeEach: function (client, done) {
         client.url('http://localhost:8124/atompm').pause(300).maximizeWindow(done);
+
+        mouse_tracking.track_mouse(client);
     },
 
     'Login': function (client) {
