+++ tests/mouse_tracking.js
@@ -1,0 +1,39 @@
+
+
+function track_mouse(client) {
+    client.execute(function() {
+        console.log("MOUSE MOUSE MOUSE");
+        // https://gist.github.com/primaryobjects/70087610d9aef0f4bddbe2101dda7649
+
+        // Create mouse following image.
+        var seleniumFollowerImg = document.createElement("img");
+
+        // Set image properties.
+        seleniumFollowerImg.setAttribute('src', 'data:image/png;base64,'
+            + 'iVBORw0KGgoAAAANSUhEUgAAABQAAAAeCAQAAACGG/bgAAAAAmJLR0QA/4ePzL8AAAAJcEhZcwAA'
+            + 'HsYAAB7GAZEt8iwAAAAHdElNRQfgAwgMIwdxU/i7AAABZklEQVQ4y43TsU4UURSH8W+XmYwkS2I0'
+            + '9CRKpKGhsvIJjG9giQmliHFZlkUIGnEF7KTiCagpsYHWhoTQaiUUxLixYZb5KAAZZhbunu7O/PKf'
+            + 'e+fcA+/pqwb4DuximEqXhT4iI8dMpBWEsWsuGYdpZFttiLSSgTvhZ1W/SvfO1CvYdV1kPghV68a3'
+            + '0zzUWZH5pBqEui7dnqlFmLoq0gxC1XfGZdoLal2kea8ahLoqKXNAJQBT2yJzwUTVt0bS6ANqy1ga'
+            + 'VCEq/oVTtjji4hQVhhnlYBH4WIJV9vlkXLm+10R8oJb79Jl1j9UdazJRGpkrmNkSF9SOz2T71s7M'
+            + 'SIfD2lmmfjGSRz3hK8l4w1P+bah/HJLN0sys2JSMZQB+jKo6KSc8vLlLn5ikzF4268Wg2+pPOWW6'
+            + 'ONcpr3PrXy9VfS473M/D7H+TLmrqsXtOGctvxvMv2oVNP+Av0uHbzbxyJaywyUjx8TlnPY2YxqkD'
+            + 'dAAAAABJRU5ErkJggg==');
+        seleniumFollowerImg.setAttribute('id', 'selenium_mouse_follower');
+        seleniumFollowerImg.setAttribute('style', 'position: absolute; z-index: 99999999999; pointer-events: none;');
+
+        // Add mouse follower to the web page.
+        document.body.appendChild(seleniumFollowerImg);
+
+        // Track mouse movements and re-position the mouse follower.
+        $(document).mousemove(function(e) {
+            $("#selenium_mouse_follower").css({ left: e.pageX, top: e.pageY });
+        })
+
+    });
+}
+
+module.exports = {
+    '@disabled': true,
+    track_mouse
+};
